# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Proteomics Visualization API
  description: Proteomics Visualization Web App API
  version: 1.0.0
paths:
...
  /api/sessions:
    get:
      tags:
      - sessions
      summary: List Sessions
      description: List all sessions.
      operationId: list_sessions_api_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SessionSummary'
                type: array
                title: Response List Sessions Api Sessions Get
    post:
      tags:
      - sessions
      summary: Create Session
      description: Create a new analysis session.
      operationId: create_session_api_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sessions/{session_id}:
    get:
      tags:
      - sessions
      summary: Get Session
      description: Get a specific session.
      operationId: get_session_api_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - sessions
      summary: Update Session
      description: Update a session.
      operationId: update_session_api_sessions__session_id__put
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sessions
      summary: Delete Session
      description: Delete a session and all its data.
      operationId: delete_session_api_sessions__session_id__delete
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FileInfo:
      properties:
        filename:
          type: string
          title: Filename
          description: Canonical filename used on disk
        original_filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Original Filename
          description: Filename supplied by the user before sanitization
        size:
          type: integer
          minimum: 0.0
          title: Size
          description: File size in bytes
        uploaded_at:
          type: string
          format: date-time
          title: Uploaded At
        columns:
          items:
            type: string
          type: array
          title: Columns
          description: CSV columns
      type: object
      required:
      - filename
      - size
      title: FileInfo
      description: File metadata model.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProteomicsFileInfo:
      properties:
        filename:
          type: string
          title: Filename
          description: Canonical filename used on disk
        original_filename:
          anyOf:
          - type: string
          - type: 'null'
          title: Original Filename
          description: Filename supplied by the user before sanitization
        size:
          type: integer
          minimum: 0.0
          title: Size
          description: File size in bytes
        uploaded_at:
          type: string
          format: date-time
          title: Uploaded At
        columns:
          items:
            type: string
          type: array
          title: Columns
          description: CSV columns
        experiment:
          type: string
          title: Experiment
          description: Experiment name (user-provided)
          default: ''
        replicate:
          type: integer
          minimum: 0.0
          title: Replicate
          description: Replicate number (user-provided)
          default: 0
        batch:
          anyOf:
          - type: string
          - type: 'null'
          title: Batch
          description: Batch label for DIA batch correction
        file_type:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
          description: 'Detected file type: ''tmt'' or ''dia'''
        tmt_channels:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tmt Channels
          description: Detected TMT reporter ion channels
        has_quan_value:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Quan Value
          description: Whether the file has a Quan Value column (DIA)
      type: object
      required:
      - filename
      - size
      title: ProteomicsFileInfo
      description: Proteomics file metadata — user-provided, not filename-parsed.
    Session:
      properties:
        id:
          type: string
          title: Id
          description: Unique session ID (UUID)
        name:
          type: string
          minLength: 1
          title: Name
          description: Session name
        template:
          type: string
          title: Template
          default: multi_condition_comparison
        pipeline:
          type: string
          title: Pipeline
          default: ''
        state:
          $ref: '#/components/schemas/SessionState'
          default: created
        config:
          anyOf:
          - $ref: '#/components/schemas/SessionConfig'
          - type: 'null'
        files:
          $ref: '#/components/schemas/SessionFiles'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        markers:
          additionalProperties:
            items:
              type: string
            type: array
          type: object
          title: Markers
          description: Marked protein accessions per comparison for volcano plot labels
        volcano_filters:
          anyOf:
          - type: object
          - type: 'null'
          title: Volcano Filters
          description: Volcano plot filter settings (foldChange, pValue, adjPValue, s0)
      type: object
      required:
      - id
      - name
      title: Session
      description: Session model representing a complete analysis session.
      example:
        created_at: '2026-03-16T10:00:00Z'
        files:
          proteomics: []
        id: 550e8400-e29b-41d4-a716-446655440000
        name: DMSO vs Treatment Analysis
        pipeline: msqrob2
        state: created
        template: multi_condition_comparison
        updated_at: '2026-03-16T10:00:00Z'
    SessionConfig:
      properties:
        treatment:
          anyOf:
          - type: string
          - type: 'null'
          title: Treatment
          description: Treatment condition name
        control:
          anyOf:
          - type: string
          - type: 'null'
          title: Control
          description: Control condition name
        organism:
          anyOf:
          - type: string
            pattern: ^(|[a-z]+)$
          - type: 'null'
          title: Organism
          description: Organism identifier (e.g., 'human', 'mouse')
        remove_razor:
          type: boolean
          title: Remove Razor
          description: Whether to remove razor peptides
          default: false
        strict_filtering:
          type: boolean
          title: Strict Filtering
          description: Use strict filtering criteria
          default: false
        comparisons:
          anyOf:
          - items:
              additionalProperties:
                additionalProperties:
                  type: string
                type: object
              type: object
            type: array
          - type: 'null'
          title: Comparisons
          description: 'List of {group1: {col:val}, group2: {col:val}} comparison criteria'
        metadata_columns:
          anyOf:
          - additionalProperties:
              additionalProperties:
                type: string
              type: object
            type: object
          - type: 'null'
          title: Metadata Columns
          description: Custom metadata columns per sample file
        msstats_normalization:
          anyOf:
          - type: string
          - type: 'null'
          title: Msstats Normalization
        msstats_feature_selection:
          anyOf:
          - type: string
          - type: 'null'
          title: Msstats Feature Selection
        msstats_summary_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Msstats Summary Method
        msstats_impute:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msstats Impute
        msstats_log_base:
          anyOf:
          - type: integer
          - type: 'null'
          title: Msstats Log Base
        msstats_censored_int:
          anyOf:
          - type: string
          - type: 'null'
          title: Msstats Censored Int
        msstats_max_quantile:
          anyOf:
          - type: number
          - type: 'null'
          title: Msstats Max Quantile
        msstats_remove50missing:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msstats Remove50Missing
        pvalue_threshold:
          anyOf:
          - type: number
            maximum: 0.5
            minimum: 0.001
          - type: 'null'
          title: Pvalue Threshold
        logfc_threshold:
          anyOf:
          - type: number
            maximum: 5.0
            minimum: 0.1
          - type: 'null'
          title: Logfc Threshold
        min_peptides_per_protein:
          anyOf:
          - type: integer
            maximum: 10.0
            minimum: 1.0
          - type: 'null'
          title: Min Peptides Per Protein
        msstats_n_top_feature:
          anyOf:
          - type: integer
          - type: 'null'
          title: Msstats N Top Feature
        msstats_min_feature_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Msstats Min Feature Count
        msstats_remove_uninformative_feature_outlier:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msstats Remove Uninformative Feature Outlier
        msstats_equal_feature_var:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msstats Equal Feature Var
        msstats_name_standards:
          anyOf:
          - type: string
          - type: 'null'
          title: Msstats Name Standards
        msstats_save_fitted_models:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msstats Save Fitted Models
        msstats_n_cores:
          anyOf:
          - type: integer
          - type: 'null'
          title: Msstats N Cores
        msqrob2_ridge:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Msqrob2 Ridge
        msqrob2_normalization:
          anyOf:
          - type: string
          - type: 'null'
          title: Msqrob2 Normalization
        msqrob2_imputation:
          anyOf:
          - type: string
          - type: 'null'
          title: Msqrob2 Imputation
        msqrob2_aggregation:
          anyOf:
          - type: string
          - type: 'null'
          title: Msqrob2 Aggregation
        msqrob2_adjust_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Msqrob2 Adjust Method
        msqrob2_n_cores:
          anyOf:
          - type: integer
          - type: 'null'
          title: Msqrob2 N Cores
        msqrob2_batch_column:
          anyOf:
          - type: string
          - type: 'null'
          title: Msqrob2 Batch Column
        covariate_columns:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Covariate Columns
        file_type:
          anyOf:
          - type: string
          - type: 'null'
          title: File Type
          description: 'Analysis type: ''tmt'' or ''dia'''
        tmt_channel_mapping:
          anyOf:
          - additionalProperties:
              additionalProperties:
                anyOf:
                - type: string
                - type: integer
              type: object
            type: object
          - type: 'null'
          title: Tmt Channel Mapping
          description: 'TMT channel → {group1: val1, ..., replicate: N} mapping'
      additionalProperties: false
      type: object
      title: SessionConfig
      description: Session configuration model.
    SessionCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        template:
          type: string
          title: Template
          default: multi_condition_comparison
        pipeline:
          type: string
          title: Pipeline
          default: msqrob2
      type: object
      required:
      - name
      title: SessionCreate
      description: Session creation request.
    SessionFiles:
      properties:
        proteomics:
          items:
            $ref: '#/components/schemas/ProteomicsFileInfo'
          type: array
          title: Proteomics
        ptm_enrichment:
          items:
            $ref: '#/components/schemas/ProteomicsFileInfo'
          type: array
          title: Ptm Enrichment
        global_proteome:
          items:
            $ref: '#/components/schemas/ProteomicsFileInfo'
          type: array
          title: Global Proteome
        fasta:
          items:
            $ref: '#/components/schemas/FileInfo'
          type: array
          title: Fasta
      type: object
      title: SessionFiles
      description: Session files collection.
    SessionState:
      type: string
      enum:
      - created
      - configuring
      - queued
      - processing
      - completed
      - error
      - cancelled
      title: SessionState
      description: Session state enumeration.
    SessionSummary:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        state:
          $ref: '#/components/schemas/SessionState'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        has_results:
          type: boolean
          title: Has Results
          default: false
      type: object
      required:
      - id
      - name
      - state
      - created_at
      - updated_at
      title: SessionSummary
      description: Session summary for list views.
    SessionUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
        config:
          anyOf:
          - $ref: '#/components/schemas/SessionConfig'
          - type: 'null'
      type: object
      title: SessionUpdate
      description: Session update request.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...